feat(Journal): transform front flyer teaser into guten tag separator
