+++ apps/publikator/components/Templates/withFrontData.js
@@ -1,10 +1,6 @@
 import { graphql } from '@apollo/client/react/hoc'
 import { gql } from '@apollo/client'
-import {
-  TeaserActiveDebates,
-  TeaserMyMagazine,
-  TeaserFlyer,
-} from '@project-r/styleguide'
+import { TeaserActiveDebates, TeaserMyMagazine } from '@project-r/styleguide'
 
 const feedQuery = gql`
   query getFrontFeed(
@@ -111,11 +107,4 @@ export const withMyMagazineData = graphql(
     ${TeaserMyMagazine.data.query}
   `,
   TeaserMyMagazine.data.config,
-)
-
-export const withFlyerData = graphql(
-  gql`
-    ${TeaserFlyer.data.query}
-  `,
-  TeaserFlyer.data.config,
 )
