+++ apps/www/components/Front/withData.js
@@ -1,10 +1,6 @@
 import { graphql } from '@apollo/client/react/hoc'
 import { gql } from '@apollo/client'
-import {
-  TeaserActiveDebates,
-  TeaserMyMagazine,
-  TeaserFlyer,
-} from '@project-r/styleguide'
+import { TeaserActiveDebates, TeaserMyMagazine } from '@project-r/styleguide'
 
 export const FRONT_FEED_QUERY = gql`
   query getFrontFeed(
@@ -117,11 +113,4 @@ export const withMyMagazineData = graphql(
     ${TeaserMyMagazine.data.query}
   `,
   TeaserMyMagazine.data.config,
-)
-
-export const withFlyerData = graphql(
-  gql`
-    ${TeaserFlyer.data.query}
-  `,
-  TeaserFlyer.data.config,
 )
