+++ packages/styleguide/src/components/SeparatorGutenTag/docs.md
@@ -1,0 +1,5 @@
+Used to separate today's articles from previous articles on the front.
+
+```react
+<SeparatorGutenTag />
+```
