+++ packages/styleguide/src/components/SeparatorGutenTag/index.js
@@ -1,0 +1,1 @@
+export { default as SeparatorGutenTag } from './SeparatorGutenTag'
