--- packages/styleguide/src/components/TeaserFlyer/index.js
@@ -1,1 +1,0 @@
-export { default as TeaserFlyer } from './TeaserFlyer'
